fixed again

+++ FVCalc.java
@@ -6,6 +6,6 @@ public class FVCalc {
 		double n = Double.parseDouble(args[2]);
       double futureValue = currentValue*Math.pow(1+(rate/100),n);
 	  System.out.println("After "+ (int)n +" years, $" + (int)currentValue +
-	  " saved at "+ (int)rate+"% will yield $" + (int)futureValue);
+	  " saved at "+ rate +"% will yield $" + (int)futureValue);
 	}
 }
